// models/parent.ts
export interface Parent {
  id:number;
  nomP: string;
  prenomP: string;
  civilite: string;
  tel1: string;
  tel2: string;
  emailP: string;
  penseBeteP:string;
  adressePostaleP: string;
  lienParente: string;
  prof: string;
  lieuNP: string;
  signiatureP: string;
  creePar:number;
  modifiePar:number;
  createdAt: string;
  updatedAt: string;
  idCU:number;
  nomE:string;
  idEcole:string;
  nom:string;
  prenom:string;
  nomC:string;
  avatar:string;
}